+++ app/src/main/java/com/laudhoot/persistence/ArchiveStatus.java
@@ -4,7 +4,7 @@ package com.laudhoot.persistence;
  * Created by apurve on 1/3/15.
  */
 public class ArchiveStatus {
-    public static final byte NOT_ARCHIVED = 0;
-    public static final byte ARCHIVED = 1;
-    public static final byte MARKED_FOR_ARCHIVES = 2;
+    public static final int NOT_ARCHIVED = 0;
+    public static final int ARCHIVED = 1;
+    public static final int MARKED_FOR_ARCHIVES = 2;
 }
